K0076_CPR code update

+++ source/Lib/CommonLib/Mv.h
@@ -273,7 +273,7 @@ namespace std
   {
     size_t operator()(const Mv& value) const
     {
-      return (((unsigned long long)value.hor << 32) + value.ver);
+      return (((uint64_t)value.hor << 32) + value.ver);
     }
   };
 };
